requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.31
Flask-Login==0.6.3
Flask-WTF==0.15.1
PyMySQL==1.0.2
pytest==6.2.4
coverage==5.5
black==21.7b0
pylint==3.2.3
bandit==1.7.0
python-dotenv
Werkzeug==3.0.1
gunicorn
Flask-SQLAlchemy==3.1.1
